add spring security and jwt token part

+++ src/main/java/com/bhavesh/airbnbapp/service/HotelServiceImpl.java
@@ -5,13 +5,16 @@ import com.bhavesh.airbnbapp.dto.HotelInfoDto;
 import com.bhavesh.airbnbapp.dto.RoomDto;
 import com.bhavesh.airbnbapp.entity.Hotel;
 import com.bhavesh.airbnbapp.entity.Room;
+import com.bhavesh.airbnbapp.entity.User;
 import com.bhavesh.airbnbapp.exception.ResourceNotFoundException;
+import com.bhavesh.airbnbapp.exception.UnAuthorisedException;
 import com.bhavesh.airbnbapp.repository.HotelRepository;
 import com.bhavesh.airbnbapp.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,6 +34,10 @@ public class HotelServiceImpl implements HotelService {
         log.info("Creating a new hotel with name: {}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with id: {}", hotel.getId());
         return modelMapper.map(hotel, HotelDto.class);
@@ -42,6 +49,12 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("Hotel does not belong to this user with id: " + user.getId());
+        }
+
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -51,6 +64,12 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("Hotel does not belong to this user with id: " + user.getId());
+        }
+
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -66,6 +85,11 @@ public class HotelServiceImpl implements HotelService {
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: " + id));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("Hotel does not belong to this user with id: " + user.getId());
+        }
 
         for (Room room : hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
@@ -82,14 +106,22 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: " + hotelId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("Hotel does not belong to this user with id: " + user.getId());
+        }
+
         hotel.setActive(true);
 
 //        assuming only do it once
-         for (Room room : hotel.getRooms()) {
+        for (Room room : hotel.getRooms()) {
             inventoryService.initializeRoomForAYear(room);
         }
     }
 
+
+    //    public method
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
         Hotel hotel = hotelRepository
